requirements.txt
Django==1.8
djangorestframework==3.1.1
pytz==2018.5
